# Web/IMBP.Web.Client/src/api/generated/models/authenticationResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * IMBP.Web.Server | v1
 * OpenAPI spec version: 1.0.0
 */

export interface AuthenticationResponse {
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  middleName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  fullName?: string | null;
}